Add the 5 missing secrets-vault migrations
Modules\Core\Models\{Secret,SecretAccessLog,SecretRotationPolicy,
SecretAccessGrant,ApiKey} have always declared their $table properties
(core_secrets, core_secret_access_logs, core_secret_rotation_policies,
core_secret_access_grants, core_api_keys), but no migration anywhere ever
created any of them — the whole vault has been dormant since it was
written, independent of the missing EncryptionService fixed in the prior
two commits.

Columns derived from each model's real $fillable/$casts, not from the
existing factories (confirmed generic/broken boilerplate — fake()->word()
for every column regardless of type, unrelated to the real schema).
Notable choices: core_secrets.encrypted_value is longText, not text — the
"SEC1.<version>.<payload>" envelope EncryptionService produces roughly
2.4x-expands a plaintext up to the 65535-byte max this module's config
allows, right at TEXT's ceiling, where truncation would silently corrupt a
stored secret; core_secret_access_logs.secret_id is nullable (an
access-denied log can reference no resolvable secret at all); unique
(tenant_id, name) on core_secrets backstops the TOCTOU-racy manual
uniqueness check in SecretsService::storeSecret().

Also fixes two bugs SecretsServiceTest surfaced only once these tables
existed and could actually be reached (previously masked behind the
missing-class/missing-table errors fixed in the prior two commits):
- SecretRotationPolicy::daysUntilRotation() declared ?int but
  now()->diffInDays() returns float in this Carbon version — a dormant
  TypeError, same class of bug as the AuditService fix two commits back.
- test_secret_expiration_enforced and test_retrieve_secret_decrypts_value
  (setUp) assumed things that don't hold: inserting the raw (null)
  $this->user->tenant_id into a NOT NULL column instead of the 'default'
  string SecretsService::getTenantId() actually falls back to, and
  retrieving a secret as a non-admin actor with no access grant
  (SecretAccessControl::canAccessSecret() only bypasses its grant check
  for admin/super-admin — storing a secret never grants the creator
  access to it). Both are test-setup fixes, no service code changed for
  either.

Result: SecretsServiceTest goes from every test failing on either the
missing class or the missing table to 12/15 passing outright; the
remaining 2 (masking returns plaintext, audit log never written) are
exactly the config-merge bug the next commit fixes — confirmed nothing
else is silently broken by comparing against the full 15-test list before
touching config.

+++ Modules/Core/app/Models/SecretRotationPolicy.php
@@ -85,7 +85,7 @@ class SecretRotationPolicy extends Model
             return null;
         }
 
-        $days = now()->diffInDays($this->next_rotation_at);
+        $days = (int) now()->diffInDays($this->next_rotation_at);
         return $days < 0 ? null : $days;
     }
